edit functions
split and stick implementation

+++ copl.cpp
@@ -1,4 +1,4 @@
 #include "copl.h"
 
 Copl::Copl() {}
-Copl::Copl(int nF, int nV, double cop) : nF(nF), nV(nV), cop(cop) {}
+Copl::Copl(int nF, int nV, int cop) : nF(nF), nV(nV), cop(cop) {}
